Commit: Add Manager service

+++ AurigainLoanERPApi/AurigainLoanERP.Database/District.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[District]
 (
-	[Id] INT NOT NULL PRIMARY KEY Identity(1,1), 
+	[Id] BIGINT NOT NULL PRIMARY KEY Identity(1,1), 
     [Name] NVARCHAR(500) NOT NULL, 
     [Pincode] nvarchar(30) NOT NULL,
     [State_Id] INT NOT NULL,
